+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+* `bootsnap precompile` CLI will now also precompile `Rakefile` and `.rake` files.
+
 * Remove `uname` and other patform specific version from the cache keys. `RUBY_PLATFORM + RUBY_REVISION` should be
   enough to ensure bytecode compatibility. This should improve caching for alpine based setups. See #409.